requirements.txt
faker
django-base64field
requests
numpy
pillow